ffi: rename `coinstr-ffi` to `coinstr-sdk-ffi`
