--- bindings/coinstr-ffi/bindings-android/lib/src/main/AndroidManifest.xml
@@ -1,6 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="io.coinstr">
-
-  <uses-permission android:name="android.permission.INTERNET" />
-
-</manifest>
