--- bindings/coinstr-ffi/bindings-swift/.gitignore
@@ -1,7 +1,0 @@
-.swiftpm/
-.build/
-*.xcodeproj
-Sources/
-**/coinstrFFI.h
-**/coinstrFFI
-**/coinstrFFI.modulemap
