--- bindings/coinstr-ffi/bindings-swift/Tests/CoinstrTests/CoinstrTests.swift
@@ -1,6 +1,0 @@
-import XCTest
-@testable import Coinstr
-
-final class CoinstrTests: XCTestCase {
-    func testExample() throws {}
-}
