--- bindings/coinstr-ffi/bindings-swift/coinstrFFI.xcframework/ios-arm64/breez_sdkFFI.framework/Headers/coinstrFFI-umbrella.h
@@ -1,1 +1,0 @@
-#import "coinstrFFI.h"
