--- bindings/coinstr-ffi/bindings-swift/coinstrFFI.xcframework/ios-arm64_x86_64-simulator/breez_sdkFFI.framework/Headers/coinstrFFI-umbrella.h
@@ -1,1 +1,0 @@
-#import "coinstrFFI.h"
