--- bindings/coinstr-ffi/bindings-swift/coinstrFFI.xcframework/ios-arm64_x86_64-simulator/breez_sdkFFI.framework/Modules/module.modulemap
@@ -1,6 +1,0 @@
-framework module coinstrFFI {
-  umbrella header "coinstrFFI-umbrella.h"
-
-  export *
-  module * { export * }
-}
