+++ bindings/coinstr-sdk-ffi/.gitignore
