+++ bindings/coinstr-sdk-ffi/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "coinstr-sdk-ffi"
+version = "0.1.0"
+edition = "2021"
+description = "Coinstr SDK FFI."
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+
+[lib]
+name = "coinstr_sdk_ffi"
+crate-type = ["cdylib", "staticlib"]
+
+[[bin]]
+name = "uniffi-bindgen"
+path = "uniffi-bindgen.rs"
+
+[dependencies]
+coinstr-sdk = { path = "../../crates/coinstr-sdk", features = ["blocking"] }
+uniffi = { version = "0.23", features = ["cli"] }
+uniffi_bindgen = "0.23"
+uniffi_macros = "0.23"
+
+[build-dependencies]
+uniffi = { version = "0.23", features = ["build"] }
