+++ bindings/coinstr-sdk-ffi/bindings-android/.gitignore
