+++ bindings/coinstr-sdk-ffi/bindings-android/gradle.properties
