+++ bindings/coinstr-sdk-ffi/bindings-android/lib/src/main/AndroidManifest.xml
@@ -1,0 +1,6 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+  package="io.coinstr.sdk">
+
+  <uses-permission android:name="android.permission.INTERNET" />
+
+</manifest>
