+++ bindings/coinstr-sdk-ffi/bindings-android/settings.gradle.kts
