+++ bindings/coinstr-sdk-ffi/bindings-swift/.gitignore
@@ -1,0 +1,7 @@
+.swiftpm/
+.build/
+*.xcodeproj
+Sources/
+**/coinstr_sdkFFI.h
+**/coinstr_sdkFFI
+**/coinstr_sdkFFI.modulemap
