+++ bindings/coinstr-sdk-ffi/bindings-swift/Tests/CoinstrSDKTests/CoinstrSDKTests.swift
@@ -1,0 +1,6 @@
+import XCTest
+@testable import CoinstrSDK
+
+final class CoinstrSDKTests: XCTestCase {
+    func testExample() throws {}
+}
