+++ bindings/coinstr-sdk-ffi/bindings-swift/coinstr_sdkFFI.xcframework/ios-arm64_x86_64-simulator/coinstr_sdkFFI.framework/Headers/coinstr_sdkFFI-umbrella.h
@@ -1,0 +1,1 @@
+#import "coinstr_sdkFFI.h"
