+++ bindings/coinstr-sdk-ffi/bindings-swift/coinstr_sdkFFI.xcframework/ios-arm64_x86_64-simulator/coinstr_sdkFFI.framework/Modules/module.modulemap
@@ -1,0 +1,6 @@
+framework module coinstr_sdkFFI {
+  umbrella header "coinstr_sdkFFI-umbrella.h"
+
+  export *
+  module * { export * }
+}
