+++ bindings/coinstr-sdk-ffi/src/client.rs
