+++ bindings/coinstr-sdk-ffi/src/error.rs
