+++ bindings/coinstr-sdk-ffi/src/proposal.rs
