+++ bindings/coinstr-sdk-ffi/uniffi.toml
@@ -1,0 +1,8 @@
+[bindings.kotlin]
+package_name = "coinstr_sdk"
+cdylib_name = "coinstr_sdk_ffi"
+
+[bindings.swift]
+ffi_module_name = "coinstr_sdk"
+ffi_module_filename = "coinstr_sdkFFI"
+cdylib_name = "coinstr_sdk_ffi"
